+++ CHANGELOG.md
@@ -7,9 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- code coverage
+
 ### Fixed
 
 - clippy lints
+- readme link to documentation
 
 ## [0.0.2] - 2023-11-22